chore: moved some components, fixed tests
