+++ src/renderer/features/operation/init/ui/MultishardStaking/MultishardStaking.tsx
