+++ src/renderer/features/operation/init/ui/MultisigOperation/MultisigOperation.tsx
