+++ src/renderer/features/operation/init/ui/OperationFooter/OperationFooter.tsx
