+++ src/renderer/features/operation/sign/index.ts
@@ -1,1 +1,2 @@
 export * from './ui/Signing/Signing';
+export * from './model';
