+++ src/renderer/features/operation/sign/ui/Signing/Signing.tsx
@@ -1,6 +1,6 @@
-import { SigningProps } from '@renderer/features/signing/model';
+import { SigningProps } from '../../model';
 import { SigningType } from '@renderer/domain/shared-kernel';
-import { VaultSigning } from '@renderer/features/signing/ui/VaultSigning/VaultSigning';
+import { VaultSigning } from '../VaultSigning/VaultSigning';
 
 export const Signing = (props: SigningProps) => {
   const signingType = props.accounts[0].signingType;
