+++ src/renderer/pages/Operations/components/modals/ApproveTx.tsx
@@ -25,7 +25,7 @@ import Confirmation from '@renderer/pages/Operations/components/ActionSteps/Conf
 import SignatorySelectModal from '@renderer/pages/Operations/components/modals/SignatorySelectModal';
 import OperationModalTitle from '@renderer/pages/Operations/components/OperationModalTitle';
 import { useMultisigEvent } from '@renderer/entities/multisig';
-import { Signing } from '@renderer/features/signing';
+import { Signing } from '@renderer/features/operation';
 import { useBalanceValidation } from '@renderer/entities/transaction/lib/useBalanceValidation';
 
 type Props = {
