+++ src/renderer/pages/Staking/Operations/ChangeValidators/ChangeValidators.test.tsx
@@ -58,23 +58,13 @@ jest.mock('./InitOperation/InitOperation', () => ({ onResult }: any) => {
 
 jest.mock('../components/index', () => ({
   Validators: ({ onResult }: any) => mockButton('to confirm', onResult),
-  Confirmation: ({ onResult }: any) => mockButton('to scan', onResult),
-  Signing: ({ onResult }: any) => mockButton('to submit', onResult),
+  Confirmation: ({ onResult }: any) => mockButton('to sign', onResult),
   Submit: () => 'finish',
 }));
 
-jest.mock(
-  '@renderer/components/common/Scanning/ScanMultiframeQr',
-  () =>
-    ({ onResult }: any) =>
-      mockButton('to sign', onResult),
-);
-jest.mock(
-  '@renderer/components/common/Scanning/ScanSingleframeQr',
-  () =>
-    ({ onResult }: any) =>
-      mockButton('to sign', onResult),
-);
+jest.mock('@renderer/features/operation', () => ({
+  Signing: ({ onResult }: any) => mockButton('to submit', onResult),
+}));
 
 describe('pages/Staking/ChangeValidators', () => {
   test('should render component', async () => {
@@ -97,9 +87,6 @@ describe('pages/Staking/ChangeValidators', () => {
     await act(async () => nextButton.click());
 
     nextButton = screen.getByRole('button', { name: 'to confirm' });
-    await act(async () => nextButton.click());
-
-    nextButton = screen.getByRole('button', { name: 'to scan' });
     await act(async () => nextButton.click());
 
     nextButton = screen.getByRole('button', { name: 'to sign' });
