+++ src/renderer/pages/Staking/Operations/Destination/Destination.tsx
@@ -14,7 +14,7 @@ import { MultisigAccount, isMultisig, Account, useAccount } from '@renderer/enti
 import { DestinationType } from '../common/types';
 import { BaseModal, Button, Loader } from '@renderer/shared/ui';
 import OperationModalTitle from '@renderer/pages/Operations/components/OperationModalTitle';
-import { Signing } from '@renderer/features/signing';
+import { Signing } from '@renderer/features/operation';
 
 const enum Step {
   INIT,
