+++ src/renderer/pages/Staking/Operations/Redeem/InitOperation/InitOperation.tsx
@@ -12,7 +12,7 @@ import { StakingMap, useStakingData, useEra } from '@renderer/entities/staking';
 import { OperationForm } from '../../components';
 import { validateBalanceForFee, validateBalanceForFeeDeposit, getRedeemAccountOption } from '../../common/utils';
 import { getSignatoryOption } from '@renderer/pages/Transfer/common/utils';
-import { OperationFooter, OperationHeader } from '@renderer/features/init-operation';
+import { OperationFooter, OperationHeader } from '@renderer/features/operation';
 
 export type RedeemResult = {
   accounts: Account[];
