+++ src/renderer/pages/Staking/Operations/Redeem/Redeem.tsx
@@ -12,7 +12,7 @@ import { getRelaychainAsset, toAddress, DEFAULT_TRANSITION } from '@renderer/sha
 import { useToggle } from '@renderer/shared/lib/hooks';
 import OperationModalTitle from '@renderer/pages/Operations/components/OperationModalTitle';
 import { BaseModal, Button, Loader } from '@renderer/shared/ui';
-import { Signing } from '@renderer/features/signing';
+import { Signing } from '@renderer/features/operation';
 
 const enum Step {
   INIT,
