+++ src/renderer/pages/Staking/Operations/Restake/InitOperation/InitOperation.tsx
@@ -17,7 +17,7 @@ import {
   validateBalanceForFeeDeposit,
 } from '../../common/utils';
 import { getSignatoryOption } from '@renderer/pages/Transfer/common/utils';
-import { OperationFooter, OperationHeader } from '@renderer/features/init-operation';
+import { OperationFooter, OperationHeader } from '@renderer/features/operation';
 
 export type RestakeResult = {
   accounts: Account[];
