+++ src/renderer/pages/Transfer/components/ActionSteps/index.ts
@@ -1,6 +1,5 @@
 import InitOperation from './InitOperation';
 import Confirmation from './Confirmation';
-import Signing from './Signing';
 import Submit from './Submit';
 
-export { InitOperation, Confirmation, Signing, Submit };
+export { InitOperation, Confirmation, Submit };
